1.50.32-RC-stage

+++ apl-core/src/main/java/com/apollocurrency/aplwallet/apl/core/transaction/messages/ChildAccountAttachment.java
@@ -43,7 +43,7 @@ public class ChildAccountAttachment extends AbstractAttachment {
     public ChildAccountAttachment(JSONObject attachmentData) {
         super(attachmentData);
         this.addressScope = AddressScope.from(((Number) attachmentData.get("addressScope")).intValue());
-        this.childCount = (short) (((Number) attachmentData.get("childCount")).intValue());
+        this.childCount = (((Number) attachmentData.get("childCount")).shortValue());
         childPublicKey = new LinkedList<>();
         List<?> keys = (List<?>) attachmentData.get("childPublicKeys");
         for (Object publicKey : keys) {
